requirements.txt
streamlit
langchain
langchain-community
langchain-openai
openai
pinecone-client
python-dotenv
PyPDF2
pypdf
tiktoken